initial commit of fetch functionality
more config chagnes, added the origin to the remots and added a branch to the config for master to point at the remote/origin

changes to the way the fetch op works, removed the origin addition from the config - this was just plain wrong

fixed output from fetchOp

updated tests

added the concept of a payload which will be the transfer object between the remote service and the client

updated way remotes are interfaced, updated documentation

added repo

updates to add the full set of blobs and trees to the client repository, updated tests to comply with this

-

renamed test class

cleaning

updates for Remote fetching

Migrating a couple final bits to teh new serialisation.

updated pom

updates towards online remote fetching

Moving the application context into a package.

remote fetching - create a custom protocol, which is very similar to the git protocol for sending its packfiles

partly working implementation of remote fetching - seems there is something not write when writing the fetched Ref objects

removed reference to a commons-io Exception that we not actually being throw??

updated documentation

updated online and offline tests with a 'two remotes' test

made changes to let tests run

change directory name of resources directory

fixed a fetch bug

moved all network IO to a new class NetoworkIO which can handles all GeoGIT networking protocol

merge operation changes

merge tests added, criss cross test and get feature tests

updated merge to return a list of object ids which are the new content ID's of the features which where changed in the merge - DiffOp did the real work

added copyright to headers of new files, fixed test for PullOp

refactored the include method in PullOp to be setRepository - more inline with git parameters, removed the need to send it the full repo path - can now simply be PROJECT_NAME/PROJECT_BRANCH

changed the call condition for pull() to not have to include a repository - by default now uses origin/master if not specified

created a test for the origin only pull request

initial steps for push op

changed fetch to mater for origin setup

added PayloadEntity to send the payload to the server; added a STATUS type to the PushResult object to reflect the response from the server (OK/CONFLICT); updated tests to test both states of STATUS

push functionality complete; extracted the fetch functionality to a until class so it can be used by end points for the push operation

fixed tests to work without locking on the file system repo - I had to run the tests in fork mode so each test gets its own VM

updated test cased to be complient with OnlineTest.java pattern for online tests

updated the test case class pattern

updated url to git origin for tests

updated test url

fixed serial usage to use the factory class